Created a deployable NFT
This is a "simple" smart contract that creates a randomized (from a total of 3 possibilities, lol) token name and deploys it on the Ethereum Rinkby Test Network. 

No customization of the NFT is taking place as of yet, however this is an active working test token. Further development will be committed expanding this smart contract.