Added metadata file directory
Added a metadata file directory that will store the data of each NFT that gets created. 

This is just a sample that was added. The image can be called upon by a directory that holds a screenshot of the final form of the character.
Right now it is just holding sample information.

Next to be done:
Figure out a way for the smart contract to call the metadata.