Added on-chain metadata
Previously data was thrown into a function to be called.

Now we give the player info (Name, number of deaths, image.. act), it is all put into the Player attributes and converted into base64. The base64 is what now holds every single data of the player, this then can be used to convert into a usable .json file that is the metadata.